add test case descriptions, rename serializer to look like other serializers.py
