--- koku/api/cloud_accounts/cloud_account_serializer.py
@@ -1,11 +1,0 @@
-"""Serializer for CloudAccount."""
-from rest_framework import serializers
-
-
-class CloudAccountSerializer(serializers.Serializer):
-    """Serializer for CloudAccount."""
-
-    name = serializers.CharField()
-    value = serializers.CharField()
-    description = serializers.CharField()
-    updated_timestamp = serializers.DateTimeField()
