+++ koku/api/cloud_accounts/serializers.py
@@ -1,0 +1,26 @@
+#
+# Copyright 2020 Red Hat, Inc.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+from rest_framework import serializers
+
+
+class CloudAccountSerializer(serializers.Serializer):
+    """Serializer for Cloud Account."""
+
+    name = serializers.CharField()
+    value = serializers.CharField()
+    description = serializers.CharField()
+    updated_timestamp = serializers.DateTimeField()
